// packages/api-client/src/generated/gateway/models/problemDetails.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * taskome-gateway
 * OpenAPI spec version: 0.1.0
 */
import type { ValidationIssue } from "./validationIssue";

export interface ProblemDetails {
  detail: string;
  errors?: ValidationIssue[] | null;
  instance: string;
  request_id: string;
  status: number;
  title: string;
  type: string;
}